Add Attribute Level Schema Override
Allow the addition of a `jsonSchema.schema` property on sequelize model attributes that will
override the type generated by the TypeMapper class.  This is important for any type that is mapped
to the TypeMapper's internal `ANY` type when using the OpenAPI strategy.

One Potential improvement that does not fit nicely into the current setup would be to throw strategy
specific errors when this override is required to produce a valid output, but unfortunately the
strategy class is only called by the TypeMapper class if we need to get a nullable version of the
type.